Muevo funciones de helpers nuevos de vuelta al modelo CatalogUpload
